# Conflicts:
#	ChronoLoop/ChronoLoop/Actions/TeleportAction.hpp
#	ChronoLoop/ChronoLoop/Particles/ParticleComponents.cpp
#	ChronoLoop/ChronoLoop/Rendering/renderer.h

+++ ChronoLoop/ChronoLoop/Input/Controller.h
@@ -23,7 +23,7 @@ namespace Epoch {
 		unsigned int mSpeedCW, mSpeedCCW, mIncreaseGestureSpeed = 5;
 
 		vec2f InitialPos = { 0,0 };
-		unsigned int gestureCnt = 0;
+		float gestureCnt = 0.0f;
 
 		friend class VIM;
 	public:
